Move to Android 28 and remove emulator startup in tests.
