+++ smartcar-auth/gradle.properties
@@ -22,3 +22,5 @@ bintrayKey=ORG_GRADLE_PROJECT_bintrayKey
 vcsRepoUrl=https://github.com/smartcar/android-sdk.git
 vcsGithubRepo=smartcar/android-sdk
 vcsGithubReadme=README.md
+
+android.enableUnitTestBinaryResources=true
